Move elasticsearch utils
